Fix CI failures across all platforms and resolve PREfast static analysis warnings
- Remove CMAKE_BUILD_WITH_INSTALL_RPATH=ON so Linux/macOS test executables
  can find Boost shared libraries via build-tree RPATH
- Fix unit test expecting 8 AUX edges (model only defines 7 with empty labels)
- Scope coverage ctest to unit tests only (-R AqualinkAutomateUnitTests) to
  avoid failure from unbuilt integration test executable
- Add autoconf/automake/libtool to macOS CI and Docker base image for
  libbacktrace vcpkg port
- Suppress PREfast false positives: C26449 in node.h (Boost.URL pct_string_view),
  C28251 in tracy_memory.cpp (intentional operator new/delete overrides)
- Fix arithmetic overflow warning in MQTT reconnect delay (1 -> 1LL)
- Use const ref for error_code in protocol_thread.cpp
- Default-initialize uninitialized members in aquarite_device.h,
  jandy_message_generator_packetprocessing.h, and menu_model.h

+++ test/unit/navigation/test_navigation_menu_model.cpp
@@ -267,7 +267,7 @@ BOOST_AUTO_TEST_CASE(TestLabelAuxListPageSelectEdgeLineNumbers)
 	for (size_t i = 0; i < 7; ++i)
 	{
 		BOOST_CHECK_EQUAL(select_edges[i]->label, "");
-		BOOST_CHECK_EQUAL(select_edges[i]->trigger_line, static_cast<int>(i + 2));
+		BOOST_CHECK_EQUAL(select_edges[i]->trigger_line, static_cast<uint8_t>(i + 2));
 	}
 
 	// All AUX Select edges should target LabelAux page
